FY-195/refactor: pylint: invalid-name
