+++ src/livegame/SubGameSession/paddle.py
@@ -4,8 +4,8 @@ from enum import Enum
 from typing import Dict, Union
 
 from pong.settings import LOGLEVEL_TRACE_ENABLE
-from livegame.precision_config import round_time, round_coord, round_speed
-from livegame.SubGameConfig import SubGameConfig
+from ..precision_config import round_time, round_coord, round_speed
+from ..subgame_config import SubGameConfig
 
 
 class Player(Enum):
