+++ src/pong/asgi.py
@@ -7,7 +7,7 @@ django.setup()
 # pylint: disable=wrong-import-position
 from django.core.asgi import get_asgi_application
 from socketcontrol.events import sio
-from friends.OnlineStatusNamespace import OnlineStatusNamespace
+from friends.online_status_namespace import OnlineStatusNamespace
 
 # pylint: enable=wrong-import-position
